Update SaveManager
- Use Path.Combine to get the active platform's path separators in paths
- Switch from constants to read-only fields so that Path.Combine may be used
  to generate folder paths
- Use interpolated strings to generate file names
- Don't use interpolated strings for static strings
- Create folders files reside in before trying to write said files
- Don't remove backup file if original doesn't exist

+++ Modules/SaveManager.cs
@@ -7,13 +7,14 @@ namespace Chino_chan.Modules
 {
     public static class SaveManager
     {
-        public const string DataFolder = "Data/";
-        public const string BackupFolder = DataFolder + "Backup/";
+        public static readonly string DataFolder = "Data";
+        public static readonly string BackupFolder = Path.Combine(DataFolder, "Backup");
         static readonly List<string> Saving = new List<string>();
 
         public static T LoadSettings<T>(string fileName)
         {
-            string originalFile = DataFolder + fileName + ".json";
+            string originalFile = Path.Combine(DataFolder, $"{fileName}.json");
+            string backupFile = Path.Combine(BackupFolder, $"{fileName}.backup.json");
             T data = default;
 
             try
@@ -25,23 +26,23 @@ namespace Chino_chan.Modules
             {
                 Logger.Log(LogType.Settings, ConsoleColor.Red, "Error", $"There was an error loading { fileName }.json! Loading backup file...");
                 Logger.Log(LogType.Settings, ConsoleColor.Red, "Error", JsonConvert.SerializeObject(e, Formatting.Indented));
-                string backupFile = BackupFolder + fileName + ".backup.json";
                 if (File.Exists(backupFile))
                 {
                     try
                     {
                         data = JsonConvert.DeserializeObject<T>(File.ReadAllText(backupFile));
 
-                        Logger.Log(LogType.Settings, ConsoleColor.Green, null, $"Backup file successfully loaded! Replacing settings with backup...");
+                        Logger.Log(LogType.Settings, ConsoleColor.Green, null, "Backup file successfully loaded! Replacing settings with backup...");
 
                         if (File.Exists(originalFile))
                             File.Delete(originalFile);
 
+                        Directory.CreateDirectory(Path.GetDirectoryName(Path.GetFullPath(originalFile)));
                         File.Copy(backupFile, originalFile);
                     }
                     catch
                     {
-                        Logger.Log(LogType.Settings, ConsoleColor.Red, "Error", $"Backup file could not be loaded!");
+                        Logger.Log(LogType.Settings, ConsoleColor.Red, "Error", "Backup file could not be loaded!");
                     }
                 }
             
@@ -56,16 +57,21 @@ namespace Chino_chan.Modules
 
             Saving.Add(fileName);
 
-            string originalFile = DataFolder + fileName + ".json";
-            string backupFile = BackupFolder + fileName + ".backup.json";
+            string originalFile = Path.Combine(DataFolder, $"{fileName}.json");
+            string backupFile = Path.Combine(BackupFolder, $"{fileName}.backup.json");
 
-            if (File.Exists(backupFile))
-                File.Delete(backupFile);
+            if (File.Exists(originalFile))
+            {
+                if (File.Exists(backupFile))
+                    File.Delete(backupFile);
 
-            File.Copy(originalFile, backupFile);
+                Directory.CreateDirectory(Path.GetDirectoryName(Path.GetFullPath(backupFile)));
+                File.Copy(originalFile, backupFile);
 
-            File.Delete(originalFile);
+                File.Delete(originalFile);
+            }
 
+            Directory.CreateDirectory(Path.GetDirectoryName(Path.GetFullPath(originalFile)));
             File.WriteAllText(originalFile, JsonConvert.SerializeObject(data, Formatting.Indented));
 
             Saving.Remove(fileName);
